# src/swagger.yml
swagger: "2.0"

info:
  title: Order Router
  version: 1.0.0

schemes:
- https
host: api.deeplink.network
basePath: /

paths:
  /order_router:
    get:
      description: Route an order
      operationId: order_router
      parameters:
        - in: query
          name: sell_symbol
          type: string
          required: true
          description: The symbol of the token to sell
        - in: query
          name: sell_ID
          type: string
          required: true
          description: The ID of the token to sell
        - in: query
          name: sell_amount
          type: number
          required: true
          description: The amount of the token to sell
        - in: query
          name: buy_symbol
          type: string
          required: true
          description: The symbol of the token to buy
        - in: query
          name: buy_ID
          type: string
          required: true
          description: The ID of the token to buy
        - in: query
          name: exchanges
          type: array
          items:
            type: string
          required: false
          description: The exchanges to include in the route. Defaults to all if not provided. One of "Uniswap_V2", "Uniswap_V3", "Sushiwap_V2", "Curve"
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              pool_graph:
                type: object
                description: A dictionary of lists mapping each node to its neighbors
              path_graph:
                type: object
                description: A dictionary of lists mapping each node to its neighbors
              routes:
                type: object
                description: A dictionary of routes, each route contains the input amount, output amount, price, gas fee, and the nodes in the route. Each route also has a key for each swap in the route, each swap contains the pool, input token, input amount, output token, output amount, price impact, price, gas fee, and description

  /order_router_split:
    get:
      description: Route an order, allowing for multiple routes via order splitting
      operationId: order_router_split
      parameters:
        - in: query
          name: sell_symbol
          type: string
          required: true
          description: The symbol of the token to sell
        - in: query
          name: sell_ID
          type: string
          required: true
          description: The ID of the token to sell
        - in: query
          name: sell_amount
          type: number
          required: true
          description: The amount of the token to sell
        - in: query
          name: buy_symbol
          type: string
          required: true
          description: The symbol of the token to buy
        - in: query
          name: buy_ID
          type: string
          required: true
          description: The ID of the token to buy
        - in: query
          name: exchanges
          type: array
          items:
            type: string
          required: false
          description: The exchanges to include in the route. Defaults to all if not provided. Current options are "Uniswap_V2", "Uniswap_V3", "Sushiwap_V2", "Curve"
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              pool_graph:
                type: object
                description: A dictionary of lists mapping each node to its neighbors
              path_graph:
                type: object
                description: A dictionary of lists mapping each node to its neighbors
              routes:
                type: object
                properties:
                  gas_fee:
                    type: number
                    description: The total gas fee for all routes
                  output_amount:
                    type: number
                    description: The total output amount for all routes
                  price:
                    type: number
                    description: The total price for all routes
                  price_impact:
                    type: number
                    description: The total price impact for all routes
                  paths:
                    type: array
                    items:
                      type: object
                      properties:
                        percent:
                          type: number
                          description: The percentage of the original order that was sent to this route
                        swap_n:
                          type: object
                          properties:
                            description:
                              type: string
                              description: A description of the swap
                            exchange:
                              type: string
                              description: The exchange for the swap
                            dangerous:
                              type: boolean
                              description: Whether the swap is dangerous (i.e., goes through a pool which contains a potentially bad token)
                            gas_fee:
                              type: number
                              description: The gas fee for the swap
                            input_amount:
                              type: number
                              description: The input amount for the swap
                            input_token:
                              type: string
                              description: The input token for the swap
                            output_amount:
                              type: number
                              description: The output amount for the swap
                            output_token:
                              type: string
                              description: The output token for the swap
                            pool:
                              type: string
                              description: The pool for the swap
                            price:
                              type: number
                              description: The price for the swap
                            price_impact:
                              type: number
                              description: The price impact for the swap
                          description:
                            A swap in the route. 'n' is the swap number, starting at 0.
                    description: A list of paths the order is to be split into.
                description: A dictionary of individual paths and metadata for the whole route, including the output amount, the overall price, the overall gas fee, and the average price impact. Each path contains the input amount, output amount, the percentage of how much the original order was sent to the path, the price, gas fee, and the nodes in the path. Each path also has a key for each swap in the path, each swap contains the pool, input token, input amount, output token, output amount, price impact, price, gas fee, and description